feat: project restructuring
