+++ crates/bytecode/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+authors = ["Dragan Rakita <dragan0rakita@gmail.com>"]
+description = "Revm Database interface"
+edition = "2021"
+keywords = ["ethereum", "evm", "revm", "no_std"]
+license = "MIT"
+name = "revm-bytecode"
+repository = "https://github.com/bluealloy/revm"
+version = "1.0.0"
+readme = "../../README.md"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[lints.rust]
+unreachable_pub = "warn"
+unused_must_use = "deny"
+rust_2018_idioms = "deny"
+
+[lints.rustdoc]
+all = "warn"
+
+[dependencies]
+# revm
+revm-primitives = { path = "../primitives", version = "9.0.1", default-features = false }
+
+bitvec = { version = "1", default-features = false, features = ["alloc"] }
+
+# Optional
+serde = { version = "1.0", default-features = false, features = [
+    "derive",
+    "rc",
+], optional = true }
+
+
+[dev-dependencies]
+
+[features]
+default = ["std"]
+std = ["serde?/std", "revm-primitives/std"]
+hashbrown = ["revm-primitives/hashbrown"]
+serde = ["dep:serde", "revm-primitives/serde"]
+serde-json = ["serde"]
