+++ crates/bytecode/src/eof/decode_helpers.rs
