+++ crates/bytecode/src/legacy/jump_map.rs
@@ -1,5 +1,5 @@
-use crate::hex;
 use bitvec::vec::BitVec;
+use revm_primitives::hex;
 use std::{fmt::Debug, sync::Arc};
 
 /// A map of valid `jump` destinations.
