+++ crates/database/interface/Cargo.toml
@@ -1,0 +1,46 @@
+[package]
+authors = ["Dragan Rakita <dragan0rakita@gmail.com>"]
+description = "Revm Database interface"
+edition = "2021"
+keywords = ["ethereum", "evm", "revm", "no_std"]
+license = "MIT"
+name = "revm-database-interface"
+repository = "https://github.com/bluealloy/revm"
+version = "1.0.0"
+readme = "../../README.md"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[lints.rust]
+unreachable_pub = "warn"
+unused_must_use = "deny"
+rust_2018_idioms = "deny"
+
+[lints.rustdoc]
+all = "warn"
+
+[dependencies]
+revm-state = { path = "../../state", version = "1.0.0", default-features = false }
+
+auto_impl = "1.2"
+
+# Optional
+serde = { version = "1.0", default-features = false, features = [
+    "derive",
+    "rc",
+], optional = true }
+
+
+[dev-dependencies]
+anyhow = "1.0.83"
+criterion = "0.5"
+indicatif = "0.17"
+rstest = "0.22.0"
+alloy-sol-types = "0.8"
+
+[features]
+default = ["std"]
+std = ["serde?/std"]
+serde = ["dep:serde"]
