+++ crates/primitives/src/lib.rs
@@ -12,7 +12,6 @@ pub mod db;
 pub mod eip7702;
 pub mod env;
 
-mod bytecode;
 mod constants;
 mod evm_wiring;
 #[cfg(any(feature = "c-kzg", feature = "kzg-rs"))]
@@ -20,7 +19,6 @@ pub mod kzg;
 pub mod precompile;
 pub mod result;
 pub mod specification;
-pub mod state;
 pub mod transaction;
 pub mod utilities;
 pub use alloy_eips::eip2930::{AccessList, AccessListItem};
@@ -29,11 +27,10 @@ pub use alloy_primitives::{
     FixedBytes, Log, LogData, TxKind, B256, I256, U256,
 };
 pub use bitvec;
-pub use bytecode::*;
 pub use constants::*;
 pub use eip7702::{
-    Authorization, AuthorizationList, Eip7702Bytecode, Eip7702DecodeError, InvalidAuthorization,
-    RecoveredAuthorization, Signature, SignedAuthorization, EIP7702_MAGIC, EIP7702_MAGIC_BYTES,
+    Authorization, AuthorizationList, InvalidAuthorization, RecoveredAuthorization, Signature,
+    SignedAuthorization,
 };
 pub use env::*;
 pub use evm_wiring::*;
@@ -53,7 +50,6 @@ pub use kzg::{EnvKzgSettings, KzgSettings};
 pub use precompile::*;
 pub use result::*;
 pub use specification::*;
-pub use state::*;
 pub use transaction::Transaction;
 pub use utilities::*;
