+++ crates/state/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+authors = ["Dragan Rakita <dragan0rakita@gmail.com>"]
+description = "Revm state types"
+edition = "2021"
+keywords = ["ethereum", "evm", "revm", "no_std"]
+license = "MIT"
+name = "revm-state"
+repository = "https://github.com/bluealloy/revm"
+version = "1.0.0"
+readme = "../../README.md"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[lints.rust]
+unreachable_pub = "warn"
+unused_must_use = "deny"
+rust_2018_idioms = "deny"
+
+[lints.rustdoc]
+all = "warn"
+
+[dependencies]
+# revm
+revm-primitives = { path = "../primitives", version = "9.0.1", default-features = false }
+revm-bytecode = { path = "../bytecode", version = "1.0.0", default-features = false }
+
+# misc
+bitflags = { version = "2.6.0", default-features = false }
+
+# Optional
+serde = { version = "1.0", default-features = false, features = [
+    "derive",
+    "rc",
+], optional = true }
+
+
+[dev-dependencies]
+anyhow = "1.0.83"
+criterion = "0.5"
+indicatif = "0.17"
+rstest = "0.22.0"
+alloy-sol-types = "0.8"
+
+[features]
+default = ["std"]
+std = [
+    "serde?/std",
+    "revm-primitives/std",
+]
+serde = ["dep:serde", "revm-primitives/serde"]
