+++ crates/state/src/account_info.rs
@@ -1,0 +1,118 @@
+use core::hash::{Hash, Hasher};
+use revm_bytecode::Bytecode;
+use revm_primitives::{B256, KECCAK_EMPTY, U256};
+
+/// AccountInfo account information.
+#[derive(Clone, Debug, Eq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct AccountInfo {
+    /// Account balance.
+    pub balance: U256,
+    /// Account nonce.
+    pub nonce: u64,
+    /// code hash,
+    pub code_hash: B256,
+    /// code: if None, `code_by_hash` will be used to fetch it if code needs to be loaded from
+    /// inside `revm`.
+    pub code: Option<Bytecode>,
+}
+
+impl Default for AccountInfo {
+    fn default() -> Self {
+        Self {
+            balance: U256::ZERO,
+            code_hash: KECCAK_EMPTY,
+            code: Some(Bytecode::default()),
+            nonce: 0,
+        }
+    }
+}
+
+impl PartialEq for AccountInfo {
+    fn eq(&self, other: &Self) -> bool {
+        self.balance == other.balance
+            && self.nonce == other.nonce
+            && self.code_hash == other.code_hash
+    }
+}
+
+impl Hash for AccountInfo {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.balance.hash(state);
+        self.nonce.hash(state);
+        self.code_hash.hash(state);
+    }
+}
+
+impl AccountInfo {
+    pub fn new(balance: U256, nonce: u64, code_hash: B256, code: Bytecode) -> Self {
+        Self {
+            balance,
+            nonce,
+            code: Some(code),
+            code_hash,
+        }
+    }
+
+    /// Returns account info without the code.
+    pub fn without_code(mut self) -> Self {
+        self.take_bytecode();
+        self
+    }
+
+    /// Returns if an account is empty.
+    ///
+    /// An account is empty if the following conditions are met.
+    /// - code hash is zero or set to the Keccak256 hash of the empty string `""`
+    /// - balance is zero
+    /// - nonce is zero
+    pub fn is_empty(&self) -> bool {
+        let code_empty = self.is_empty_code_hash() || self.code_hash.is_zero();
+        code_empty && self.balance.is_zero() && self.nonce == 0
+    }
+
+    /// Returns `true` if the account is not empty.
+    pub fn exists(&self) -> bool {
+        !self.is_empty()
+    }
+
+    /// Returns `true` if account has no nonce and code.
+    pub fn has_no_code_and_nonce(&self) -> bool {
+        self.is_empty_code_hash() && self.nonce == 0
+    }
+
+    /// Return bytecode hash associated with this account.
+    /// If account does not have code, it returns `KECCAK_EMPTY` hash.
+    pub fn code_hash(&self) -> B256 {
+        self.code_hash
+    }
+
+    /// Returns true if the code hash is the Keccak256 hash of the empty string `""`.
+    #[inline]
+    pub fn is_empty_code_hash(&self) -> bool {
+        self.code_hash == KECCAK_EMPTY
+    }
+
+    /// Take bytecode from account. Code will be set to None.
+    pub fn take_bytecode(&mut self) -> Option<Bytecode> {
+        self.code.take()
+    }
+
+    pub fn from_balance(balance: U256) -> Self {
+        AccountInfo {
+            balance,
+            ..Default::default()
+        }
+    }
+
+    pub fn from_bytecode(bytecode: Bytecode) -> Self {
+        let hash = bytecode.hash_slow();
+
+        AccountInfo {
+            balance: U256::ZERO,
+            nonce: 1,
+            code: Some(bytecode),
+            code_hash: hash,
+        }
+    }
+}
