+++ examples/split_db_trait.rs
@@ -1,0 +1,207 @@
+
+
+//! Database that is split on State and BlockHash traits.
+pub mod block_hash;
+pub mod state;
+
+pub use block_hash::{BlockHash, BlockHashRef};
+pub use state::{State, StateRef};
+
+use crate::{
+    db::{Database, DatabaseRef},
+    Account, AccountInfo, Address, Bytecode, HashMap, B256, U256,
+};
+
+use super::DatabaseCommit;
+
+#[derive(Debug)]
+pub struct DatabaseComponents<S, BH> {
+    pub state: S,
+    pub block_hash: BH,
+}
+
+#[derive(Debug)]
+pub enum DatabaseComponentError<SE, BHE> {
+    State(SE),
+    BlockHash(BHE),
+}
+
+impl<S: State, BH: BlockHash> Database for DatabaseComponents<S, BH> {
+    type Error = DatabaseComponentError<S::Error, BH::Error>;
+
+    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
+        self.state.basic(address).map_err(Self::Error::State)
+    }
+
+    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error> {
+        self.state
+            .code_by_hash(code_hash)
+            .map_err(Self::Error::State)
+    }
+
+    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error> {
+        self.state
+            .storage(address, index)
+            .map_err(Self::Error::State)
+    }
+
+    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error> {
+        self.block_hash
+            .block_hash(number)
+            .map_err(Self::Error::BlockHash)
+    }
+}
+
+impl<S: StateRef, BH: BlockHashRef> DatabaseRef for DatabaseComponents<S, BH> {
+    type Error = DatabaseComponentError<S::Error, BH::Error>;
+
+    fn basic_ref(&self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
+        self.state.basic(address).map_err(Self::Error::State)
+    }
+
+    fn code_by_hash_ref(&self, code_hash: B256) -> Result<Bytecode, Self::Error> {
+        self.state
+            .code_by_hash(code_hash)
+            .map_err(Self::Error::State)
+    }
+
+    fn storage_ref(&self, address: Address, index: U256) -> Result<U256, Self::Error> {
+        self.state
+            .storage(address, index)
+            .map_err(Self::Error::State)
+    }
+
+    fn block_hash_ref(&self, number: u64) -> Result<B256, Self::Error> {
+        self.block_hash
+            .block_hash(number)
+            .map_err(Self::Error::BlockHash)
+    }
+}
+
+impl<S: DatabaseCommit, BH: BlockHashRef> DatabaseCommit for DatabaseComponents<S, BH> {
+    fn commit(&mut self, changes: HashMap<Address, Account>) {
+        self.state.commit(changes);
+    }
+}
+
+
+//! BlockHash database component from [`crate::db::Database`]
+//! it is used inside [`crate::db::DatabaseComponents`]
+
+use crate::B256;
+use auto_impl::auto_impl;
+use core::ops::Deref;
+use std::sync::Arc;
+
+#[auto_impl(&mut, Box)]
+pub trait BlockHash {
+    type Error;
+
+    /// Get block hash by block number
+    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error>;
+}
+
+#[auto_impl(&, &mut, Box, Rc, Arc)]
+pub trait BlockHashRef {
+    type Error;
+
+    /// Get block hash by block number
+    fn block_hash(&self, number: u64) -> Result<B256, Self::Error>;
+}
+
+impl<T> BlockHash for &T
+where
+    T: BlockHashRef,
+{
+    type Error = <T as BlockHashRef>::Error;
+
+    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error> {
+        BlockHashRef::block_hash(*self, number)
+    }
+}
+
+impl<T> BlockHash for Arc<T>
+where
+    T: BlockHashRef,
+{
+    type Error = <T as BlockHashRef>::Error;
+
+    fn block_hash(&mut self, number: u64) -> Result<B256, Self::Error> {
+        self.deref().block_hash(number)
+    }
+}
+
+
+//! State database component from [`crate::db::Database`]
+//! it is used inside [`crate::db::DatabaseComponents`]
+
+use crate::{AccountInfo, Address, Bytecode, B256, U256};
+use auto_impl::auto_impl;
+use core::ops::Deref;
+use std::sync::Arc;
+
+#[auto_impl(&mut, Box)]
+pub trait State {
+    type Error;
+
+    /// Get basic account information.
+    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error>;
+
+    /// Get account code by its hash
+    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error>;
+
+    /// Get storage value of address at index.
+    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error>;
+}
+
+#[auto_impl(&, &mut, Box, Rc, Arc)]
+pub trait StateRef {
+    type Error;
+
+    /// Get basic account information.
+    fn basic(&self, address: Address) -> Result<Option<AccountInfo>, Self::Error>;
+
+    /// Get account code by its hash
+    fn code_by_hash(&self, code_hash: B256) -> Result<Bytecode, Self::Error>;
+
+    /// Get storage value of address at index.
+    fn storage(&self, address: Address, index: U256) -> Result<U256, Self::Error>;
+}
+
+impl<T> State for &T
+where
+    T: StateRef,
+{
+    type Error = <T as StateRef>::Error;
+
+    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
+        StateRef::basic(*self, address)
+    }
+
+    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error> {
+        StateRef::code_by_hash(*self, code_hash)
+    }
+
+    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error> {
+        StateRef::storage(*self, address, index)
+    }
+}
+
+impl<T> State for Arc<T>
+where
+    T: StateRef,
+{
+    type Error = <T as StateRef>::Error;
+
+    fn basic(&mut self, address: Address) -> Result<Option<AccountInfo>, Self::Error> {
+        self.deref().basic(address)
+    }
+
+    fn code_by_hash(&mut self, code_hash: B256) -> Result<Bytecode, Self::Error> {
+        self.deref().code_by_hash(code_hash)
+    }
+
+    fn storage(&mut self, address: Address, index: U256) -> Result<U256, Self::Error> {
+        self.deref().storage(address, index)
+    }
+}
